Fix test which was hanging forever
This fixes a regression in our test suite caused by #1937. For some
reason, calling `insert_sean_and_tess_into_users_table` twice in the
same test causes the test to hang forever (along with any other test
running concurrently). The place where it hangs is on the insert
statement.

This fixes the issue by splitting it into two tests. The issue could
also be fixed by adding an explicit `mem::drop(connection)` on the line
before the second one is established.

I have absolutely no idea why this was causing the suite to hang like
this. The inserts should not cause a lock, so the transactions should be
able to run concurrently. If it were a deadlock issue, the test would
fail after 10 seconds, when the deadlock timeout occurs.