Remove all deprecated code
This removes all items marked `#[deprecated]` and all code with
`#[cfg(feature = "with-deprecated")]`. Unfortunately the compiler
wouldn't warn in all cases of deprecated items, particularly deprecated
re-exports. This hit us extremely hard in the case of the deprecated
re-export of `Queryable` in `prelude`, and `#[derive(Queryable)]` would
actually generate invalid code if the `with-deprecated` feature was
missing.

While I did change a lot of our code to work with the re-export of
`Queryable` removed, I wasn't actually able to remove it since
`diesel_infer_schema` breaks. We can remove it once we're no longer
using `infer_schema`, but we should also consider whether that
deprecation is worthwhile for the problems it caused.