Add `DatabaseErrorKind::ReadOnlyTransaction`
This is an error that applications are likely to want to handle. A
database may be read only during maintenance, or for brief periods after
failing over to a hot replica.

Applications designed to handle this may wish to transform this error
into a user visible HTTP 503 response, or skip certain optional writes.
By providing this error variant, we allow them to gracefully do so,
without having to resort to parsing error messages as was needed in
https://github.com/rust-lang/crates.io/pull/1670